requirements.txt
streamlit
streamlit_chat
python-dotenv
langchain
openai